Updated to use correct object
Hi Team,

While implementing typescript in my project I kept running into errors when trying to access _uid on story.

After a fair bit of digging I realised that on the live update event you are currently trying to access ui on the story object when it actually exists only on the story.content.

The code works because both return undefined so match.

I thought I would add a PR to hopefully prevent anyone else getting confused in the future.